// apps/proyecto-integrado/src/app/state/actions/backoffice/back-office-action.types.enum.ts
export enum BackOfficeActionTypes {
  FIND_USERS_REQUEST = '[Backoffice] Find Users Request',
  FIND_USERS_SUCCESS = '[Backoffice] Find Users Success',
  FIND_USERS_FAILURE = '[Backoffice] Find Users Failure',

  ADD_USER_REQUEST = '[Admin] ADD_USER_REQUEST',
  ADD_USER_SUCCESS = '[Admin] ADD_USER_SUCCESS',
  ADD_USER_FAILURE = '[Admin] ADD_USER_FAILURE',

  DELETE_USER_REQUEST = '[Admin] DELETE_USER_REQUEST',
  DELETE_USER_SUCCESS = '[Admin] DELETE_USER_SUCCESS',
  DELETE_USER_FAILURE = '[Admin] DELETE_USER_FAILURE',

  EDIT_USER_REQUEST = '[Admin] EDIT_USER_REQUEST',
  EDIT_USER_SUCCESS = '[Admin] EDIT_USER_SUCCESS',
  EDIT_USER_FAILURE = '[Admin] EDIT_USER_FAILURE',

  ADD_PARTY_REQUEST = '[Admin] ADD_PARTY_REQUEST',
  ADD_PARTY_SUCCESS = '[Admin] ADD_PARTY_SUCCESS',
  ADD_PARTY_FAILURE = '[Admin] ADD_PARTY_FAILURE',

  DELETE_PARTY_REQUEST = '[Admin] DELETE_PARTY_REQUEST',
  DELETE_PARTY_SUCCESS = '[Admin] DELETE_PARTY_SUCCESS',
  DELETE_PARTY_FAILURE = '[Admin] DELETE_PARTY_FAILURE',

  EDIT_PARTY_REQUEST = '[Admin] EDIT_PARTY_REQUEST',
  EDIT_PARTY_SUCCESS = '[Admin] EDIT_PARTY_SUCCESS',
  EDIT_PARTY_FAILURE = '[Admin] EDIT_PARTY_FAILURE',

  ADD_HOTEL_REQUEST = '[Admin] ADD_HOTEL_REQUEST',
  ADD_HOTEL_SUCCESS = '[Admin] ADD_HOTEL_SUCCESS',
  ADD_HOTEL_FAILURE = '[Admin] ADD_HOTEL_FAILURE',

  DELETE_HOTEL_REQUEST = '[Admin] DELETE_HOTEL_REQUEST',
  DELETE_HOTEL_SUCCESS = '[Admin] DELETE_HOTEL_SUCCESS',
  DELETE_HOTEL_FAILURE = '[Admin] DELETE_HOTEL_FAILURE',

  EDIT_HOTEL_REQUEST = '[Admin] EDIT_HOTEL_REQUEST',
  EDIT_HOTEL_SUCCESS = '[Admin] EDIT_HOTEL_SUCCESS',
  EDIT_HOTEL_FAILURE = '[Admin] EDIT_HOTEL_FAILURE',

  ADD_DESTINATION_REQUEST = '[Admin] ADD_DESTINATION_REQUEST',
  ADD_DESTINATION_SUCCESS = '[Admin] ADD_DESTINATION_SUCCESS',
  ADD_DESTINATION_FAILURE = '[Admin] ADD_DESTINATION_FAILURE',

  DELETE_DESTINATION_REQUEST = '[Admin] DELETE_DESTINATION_REQUEST',
  DELETE_DESTINATION_SUCCESS = '[Admin] DELETE_DESTINATION_SUCCESS',
  DELETE_DESTINATION_FAILURE = '[Admin] DELETE_DESTINATION_FAILURE',

  EDIT_DESTINATION_REQUEST = '[Admin] EDIT_DESTINATION_REQUEST',
  EDIT_DESTINATION_SUCCESS = '[Admin] EDIT_DESTINATION_SUCCESS',
  EDIT_DESTINATION_FAILURE = '[Admin] EDIT_DESTINATION_FAILURE',

  ADD_TRANSPORT_REQUEST = '[Admin] ADD_TRANSPORT_REQUEST',
  ADD_TRANSPORT_SUCCESS = '[Admin] ADD_TRANSPORT_SUCCESS',
  ADD_TRANSPORT_FAILURE = '[Admin] ADD_TRANSPORT_FAILURE',

  DELETE_TRANSPORT_REQUEST = '[Admin] DELETE_TRANSPORT_REQUEST',
  DELETE_TRANSPORT_SUCCESS = '[Admin] DELETE_TRANSPORT_SUCCESS',
  DELETE_TRANSPORT_FAILURE = '[Admin] DELETE_TRANSPORT_FAILURE',

  EDIT_TRANSPORT_REQUEST = '[Admin] EDIT_TRANSPORT_REQUEST',
  EDIT_TRANSPORT_SUCCESS = '[Admin] EDIT_TRANSPORT_SUCCESS',
  EDIT_TRANSPORT_FAILURE = '[Admin] EDIT_TRANSPORT_FAILURE',
}